Check sub-delay instead of relying on retries for extraction
Moving signs can result in querying the same
sub-text, sub-start and sub-end a ton of times, which makes it look like
the end was reached.

Increasing the retry count helps but isn't a good solution.

sub-delay changes with each step, so relying on that should make the
extraction more robust without requiring a retry counter.

Fixes #2