Close the menu on file boundaries
It doesn't make sense to keep displaying lines when the file ends.

Alternatively new lines could get fetched automatically on file change,
but then there is the problem that the file needs to be unpaused for at
least a little bit to get more then the current subtitle line, so that
won't work all that well.