Optimize first subtitle line discovery
Stepping back through subtitle lines can disrupt playback, which is a
problem for long videos with lots of lines, where stepping back to the
first line can take >1s.

Instead it's possible to set a huge subtitle delay, so big that it's
certainly before the first subtitle line, and stepping forward from
there then lands at the first line. No need to step through all the other
ones.

That makes that while loop unnecessary, but it'll stay just in case
there is a subtitle fine extreme enough for this to matter. Keeping it
around doesn't hurt.