Improve wrongly reported subtitle timing
Relying solely on sub-start and sub-end has the problem that those
reflect the earliest start time and the latest end time of the currentl
lines.

This attempts to get more accurate start and end times in cases where
sub-start and start-end doesn't correctly reflect those times of all
current lines.