Fix rare crash and simplification
I've only managed to produce the crash twice by accident, but apparently
it's possible that sometimes at least one of the values passed to
same_time() in this condition is nil.

Both times the menu was opened while subtitles for the current position
were being loaded, so it seems to be a race condition for that specific
case.