Preformat timespan for the hint
For a 5h 16min podcast the time formatting costs close to 100ms for each
update, so doing it once and remembering the result speeds up updates.

Converting to and from json costs ~270ms, so there is only so there
is still a substantial delay for updates, but it's a noticable
improvement.

Converting the menu data into menu items on the uosc side costs another
40ms, which might be improved by reusing objects from the json
deserialization.

The only way to make it truely fast would be with a new API for uosc to
allow for partial updates.